3 Link Planar Arm workspace plotting_ MATLAB
The code pulls the data points from the Wspace_mod function, with inputs here on Link lengths and respective joint actuation domains. The output array from Wspace_mod is an array CA containing workspace span coordinates at the input resolution and the column vector maxmin containing the boundary coordinates of the span which is then simulated with the joints in the code below. 

Personally, I would recommend utilising resolutions in the range of 10<res<30.  Any more will absolutely work at the cost of more time, but this code hasn't been designed for that degree of analysis and just for mere eye level observation. Please do feel free to contact me with doubts or errors at "jeano5326@gmail.com" I will be more than glad to hear any input on these.  Coded in/for MATLAB R2021a.

https://www.linkedin.com/in/jeanojoseph/
https://www.behance.net/jeanojoseph
https://github.com/jeanbuntu
@ - jeano5326@gmail.com